abstracted backup selection
